+++ src/meow/ui/widget/keyboard/KeyboardRow.cpp
@@ -166,7 +166,7 @@ namespace meow
         if (!_is_changed)
         {
             if (_visibility != INVISIBLE && _is_enabled)
-                for (uint16_t i{0}; i < _widgets.size(); ++i)
+                for (size_t i{0}; i < _widgets.size(); ++i)
                     _widgets[i]->onDraw();
         }
         else
@@ -198,7 +198,7 @@ namespace meow
             uint16_t x = step;
             uint16_t y = (float)(_height - _btn_height) / 2;
 
-            for (uint16_t i{0}; i < _widgets.size(); ++i)
+            for (size_t i{0}; i < _widgets.size(); ++i)
             {
                 _widgets[i]->setPos(x, y);
                 _widgets[i]->setWidth(_btn_width);
